components for Product and ProductDetails
